Extract and unify container runtime config generation
